+++ frontend/src/components/common/CommentInputModal/CommentInputModal.tsx
@@ -4,10 +4,10 @@ import reactDom from 'react-dom';
 import { postImageUrlConverter } from '@/api/image';
 import AnonymouseCheckBox from '@/components/common/AnonymousCheckBox/AnonymouseCheckBox';
 import * as S from '@/components/common/CommentInputModal/CommentInputModal.styles';
-import usePostCommentInputModal from '@/components/common/CommentInputModal/hooks/usePostCommentInputModal';
-import usePutCommentInputModal from '@/components/common/CommentInputModal/hooks/usePutCommentInputModal';
 import ToastUiEditor from '@/components/common/ToastUiEditor/ToastUiEditor';
-import useSnackBar from '@/hooks/useSnackBar';
+import usePostCommentInputModal from '@/hooks/comment/usePostCommentInputModal';
+import usePutCommentInputModal from '@/hooks/comment/usePutCommentInputModal';
+import useSnackBar from '@/hooks/common/useSnackBar';
 import { queryClient } from '@/index';
 import { validatedCommentInput } from '@/utils/validateInput';
 import { Editor } from '@toast-ui/react-editor';
