+++ frontend/src/components/helper/PrivateRouter.tsx
@@ -1,11 +1,10 @@
 import { useEffect } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
 
-import useSnackBar from '@/hooks/useSnackBar';
-
+import useSnackBar from '@/hooks/common/useSnackBar';
 
 const PrivateRouter = ({ isAuthenticated }: { isAuthenticated: boolean }) => {
-	const {showSnackBar} = useSnackBar();
+	const { showSnackBar } = useSnackBar();
 	useEffect(() => {
 		if (!isAuthenticated) {
 			showSnackBar('로그인 후 사용해주세요');
