+++ frontend/src/components/helper/PublicRouter.tsx
@@ -1,10 +1,10 @@
 import { useEffect } from 'react';
 import { Navigate, Outlet } from 'react-router-dom';
 
-import useSnackBar from '@/hooks/useSnackBar';
+import useSnackBar from '@/hooks/common/useSnackBar';
 
 const PublicRouter = ({ isAuthenticated }: { isAuthenticated: boolean }) => {
-	const {showSnackBar} = useSnackBar();
+	const { showSnackBar } = useSnackBar();
 	useEffect(() => {
 		if (isAuthenticated) {
 			showSnackBar('로그인 상태에서 이용할수 없는 서비스입니다');
