+++ frontend/src/pages/CategoryArticles/CategoryArticles.tsx
@@ -4,8 +4,8 @@ import ArticleItem from '@/components/common/ArticleItem/ArticleItem';
 import InfiniteScrollObserver from '@/components/common/InfiniteScrollObserver/InfiniteScrollObserver';
 import Loading from '@/components/common/Loading/Loading';
 import SortDropdown from '@/components/common/SortDropdown/SortDropDown';
+import useGetCategoryArticles from '@/hooks/article/useGetCategoryArticles';
 import * as S from '@/pages/CategoryArticles/CategoryArticles.styles';
-import useGetCategoryArticles from '@/pages/CategoryArticles/hooks/useGetCategoryArticles';
 
 const CategoryArticles = () => {
 	const navigate = useNavigate();
