+++ frontend/src/pages/Discussion/Vote/Vote.tsx
@@ -1,9 +1,9 @@
 import { MdOutlineHowToVote } from 'react-icons/md';
 
 import Loading from '@/components/common/Loading/Loading';
+import useGetVote from '@/hooks/vote/useGetVote';
 import * as S from '@/pages/Discussion/Vote/Vote.styles';
 import VoteItem from '@/pages/Discussion/VoteItem/VoteItem';
-import useGetVote from '@/pages/Discussion/hooks/useGetVote';
 
 const Vote = ({ articleId }: { articleId: string }) => {
 	const { data, isLoading, totalCount } = useGetVote(articleId);
