+++ frontend/src/pages/DiscussionDetail/index.tsx
@@ -1,8 +1,8 @@
 import { useNavigate, useParams } from 'react-router-dom';
 
 import Loading from '@/components/common/Loading/Loading';
-import useGetDetailArticle from '@/hooks/useGetDetailArticle';
-import useGetDetailComment from '@/hooks/useGetDetailComment';
+import useGetDetailArticle from '@/hooks/article/useGetDetailArticle';
+import useGetDetailComment from '@/hooks/comment/useGetDetailComment';
 import Detail from '@/pages/Detail';
 import Vote from '@/pages/Discussion/Vote/Vote';
 import VoteGenerateButton from '@/pages/DiscussionDetail/VoteGenerateButton/VoteGenerateButton';
