+++ frontend/src/pages/Login/LoginController/LoginController.tsx
@@ -4,7 +4,7 @@ import { useMutation } from 'react-query';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
 import { postLogin } from '@/api/login';
-import useSnackBar from '@/hooks/useSnackBar';
+import useSnackBar from '@/hooks/common/useSnackBar';
 
 const LoginController = () => {
 	const [searchParams] = useSearchParams();
@@ -12,10 +12,7 @@ const LoginController = () => {
 	const navigate = useNavigate();
 
 	const code = searchParams.get('code');
-	if (code === null) {
-		showSnackBar('깃허브 로그인에 동의해주세요.');
-		navigate('/');
-	}
+
 	const { data, isError, isSuccess, error, mutate } = useMutation<
 		AxiosResponse<{ accessToken: string }>,
 		AxiosError<{ errorCode: string; message: string }>,
@@ -23,6 +20,11 @@ const LoginController = () => {
 	>(postLogin);
 
 	useEffect(() => {
+		if (code === null) {
+			showSnackBar('깃허브 로그인에 동의해주세요.');
+			navigate('/');
+			return;
+		}
 		mutate(code);
 	}, []);
