+++ frontend/src/pages/UpdateWriting/index.tsx
@@ -4,7 +4,7 @@ import HashTag from '@/components/common/HashTag/HashTag';
 import Loading from '@/components/common/Loading/Loading';
 import ToastUiEditor from '@/components/common/ToastUiEditor/ToastUiEditor';
 import PageLayout from '@/components/layout/PageLayout/PageLayout';
-import usePostWritingArticle from '@/pages/UpdateWriting/hooks/usePostUpdateWritingArticle';
+import usePostWritingArticle from '@/hooks/article/usePostUpdateWritingArticle';
 import * as S from '@/pages/WritingArticles/index.styles';
 
 const UpdateWriting = () => {
